requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
Faker==19.13.0
gunicorn==21.2.0
jsonfaker==0.0.1
packaging==23.2
python-dateutil==2.8.2
pytz==2023.3.post1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
whitenoise==6.6.0
